# contracts/openapi/platform-api.yaml
openapi: 3.1.0
info:
  title: EdgeAI Control Plane
  version: 0.1.0
  description: Profile → Device/Node → Workflow DAG → Run/Task/Attempt와 검증된 결과 조회 API입니다.
    실제 Kubernetes 실행은 서버의 runtime 설정이 활성화된 환경에서 동작합니다. 쓰기 요청은 실제 DB에 반영되고 Swagger의 CSRF 토큰/쿠키는 자동 연결됩니다.
    복원 DB의 조회 전용 점검 모드에서는 실행 기능이 비활성화되고 CSRF가 유효한 쓰기 요청도 403으로 거절됩니다.
    일반 관리 API의 변경 요청은 먼저 감사 접수 기록을 저장하며 X-EdgeAI-Audit-Id를 반환합니다. 접수 저장 실패는503/AUDIT_STORE_UNAVAILABLE이며 변경을 실행하지 않습니다.
    감사의 HTTP 결과는 비동기 작업 성공을 뜻하지 않으며, 결과가 없는 접수 기록은 OUTCOME_UNKNOWN입니다. 복구 점검 모드에서는 감사 쓰기도 비활성화됩니다.
servers:
- url: /
security: []
paths:
...
  /api/v1/devices/{deviceId}/attachments/{nodeId}:
    parameters:
    - in: path
      name: deviceId
      required: true
      schema:
        type: string
        format: uuid
    - in: path
      name: nodeId
      required: true
      schema:
        type: string
        format: uuid
    put:
      tags:
      - Device
      operationId: attachDevice
      summary: 장치를 관측된 실행 노드에 연결
      description: 최근 60초 안에 실제 관측된 Ready Node와 port를 연결합니다. 기존 활성 연결은 닫고 새 이력을 만들며 동일 node/port 재전송은 이력을 추가하지 않습니다. 장치당 활성
        연결은 하나입니다.
      responses:
        '200':
          description: 요청 성공. 재전송은 기존 객체를 반환합니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceAttachment'
        '400':
          description: 필드·UUID·페이지·JSON 형식을 확인하세요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 장치·노드·참조 Profile이 없습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          description: DB 연결 장애. 복구 후 재시도하세요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: CSRF 토큰·세션 쿠키 누락/불일치
        '409':
          description: 현재 상태·버전·세션·sequence 충돌. 상세 조회 후 요청을 확인하세요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '413':
          description: 요청 크기 제한을 초과했습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
      - csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceAttachmentInput'
...
  /api/v1/virtual-devices/{vdId}/execution:
    get:
      tags: [VD]
      operationId: getVirtualDeviceExecution
      summary: VD 실행 준비 상태·세대·작업 이력 조회
      description: >-
        asOf 시점의 일관된 DB 스냅샷입니다. ready는 RUNNING 의도·Pod Ready 관측·미만료 인증 lease를 모두 요구합니다.
        과거 observedState=READY만으로 현재 준비 상태를 판단하지 마세요. current는 미종료 실행이며 없으면 null입니다.
        pendingOperation은 현재 진행 작업이며 runtimeHistory/bindings/operations는 각각 최신100개입니다.
        각 Truncated 값이 true이면 이전 이력이 더 있습니다. 기능 비활성·VD 해제 후에도 이력은 조회 가능합니다.
        세션·자격·내부 설정은 반환하지 않습니다. 성공한 Operation도 현재 실행이나 Task/Result 성공을 보장하지 않습니다.
      parameters:
      - {name: vdId, in: path, required: true, schema: {type: string, format: uuid}, description: 대상 VD UUID}
      responses:
        '200':
          description: VD 실행 상태와 보존된 이력
          content:
            application/json:
              schema: {$ref: '#/components/schemas/VDExecution'}
        '400': {$ref: '#/components/responses/VirtualDeviceError'}
        '404': {$ref: '#/components/responses/VirtualDeviceError'}
        '503': {$ref: '#/components/responses/VirtualDeviceError'}
...
  /api/v1/sensors/registration-options:
    get:
      operationId: getSensorRegistrationOptions
      summary: 배포된 수집기의 지원 센서 등록 항목 조회
      tags: [센서]
      responses:
        '200':
          description: 현재 EdgeX 서비스와 프로필이 확인된 지원 항목. 연결 가능 여부나 실측 수신을 보장하지 않는다.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SensorRegistrationCatalog'}
        '502': {description: EdgeX 목록 조회 실패 또는 조회 한도 초과}
        '503': {description: EdgeX 비활성화}
  /api/v1/sensors/registrations:
    post:
      operationId: registerSensor
      summary: 지원하는 센서를 EdgeX에 수동 등록
      tags: [센서]
      security: [{csrfToken: []}]
      description: >-
        배포된 Arduino Serial 또는 Sense HAT I2C 프로필을 사용한다. 새 드라이버나 프로필을 설치하지 않는다.
        동일 이름·연결 정보는 기존 등록을 반환하고 이름 또는 동일 포트의 측정 항목 충돌은409다.
        등록은 EdgeX DB에 저장되며 레포 YAML에는 자동 반영되지 않는다. 결과 미확인 시 자동 재시도하지 않는다.
        같은 이름·입력값을 유지한 재요청으로 확인할 수 있다. 등록 완료는 실측 수신 확인과 다르다.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SensorRegistrationRequest'}
      responses:
        '201':
          description: 새 등록 저장과 읽기 확인 완료
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SensorRegistrationResult'}
        '200':
          description: 같은 이름·설정으로 이미 등록됨
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SensorRegistrationResult'}
        '400': {description: 지원하지 않는 센서 또는 연결 설정}
        '403': {description: CSRF 검증 실패}
        '409': {description: 등록 충돌 또는 수집기·프로필 없음}
        '502': {description: EdgeX 조회 실패 또는 등록 결과 미확인}
        '503': {description: EdgeX 비활성화}
  /api/v1/sensors/readings:
    get:
      operationId: getSensorReadings
      summary: EdgeX 센서의 최근 측정값 조회
      tags: [센서]
      parameters:
      - {name: device, in: query, required: true, schema: {type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9._-]{0,127}$'}}
      - {name: resource, in: query, schema: {type: string}, description: 특정 측정 항목으로 제한}
      - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 500, default: 100}}
      responses:
        '200':
          description: 최근 N개 측정값을 측정 시각 오름차순으로 반환. 미수집은 빈 배열이며 0과 구분한다.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SensorReadings'}
        '400': {description: 잘못된 조회 조건}
        '404': {description: 센서 없음}
        '502': {description: EdgeX 조회 실패}
        '503': {description: EdgeX 비활성화}
  /api/v1/sensors/commands:
    get:
      operationId: getSensorCommands
      summary: 센서가 실제 지원하는 읽기·쓰기 명령 조회
      tags: [센서]
      parameters:
      - {name: device, in: query, required: true, schema: {type: string}}
      responses:
        '200':
          description: Core Command에서 조회한 명령. 외부 URL이나 프로토콜 인증 정보는 반환하지 않는다.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SensorCommands'}
        '404': {description: 센서 없음}
        '502': {description: EdgeX 조회 실패}
        '503': {description: EdgeX 비활성화}
  /api/v1/sensors/command:
    post:
      operationId: executeSensorCommand
      summary: EdgeX에 등록된 센서 명령 실행
      tags: [센서]
      description: CSRF가 필요하다. 명령 지원 여부와 adminState를 실행 전에 재조회한다. GET은 ds-pushevent=false로 직접 읽으며 PUT은 지정된 항목의 문자열 값을 전달한다. 자동 재시도하지 않으며 시간 초과 시 실제 장치 결과를 확인해야 한다.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SensorCommandRequest'}
      responses:
        '200':
          description: EdgeX 명령 응답. PUT 응답은 물리 장치 변경 검증을 뜻하지 않는다.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SensorCommandResult'}
        '400': {description: 지원하지 않는 명령 또는 잘못된 값}
        '403': {description: CSRF 검증 실패}
        '404': {description: 센서 또는 명령 없음}
        '409': {description: 장치 잠김}
        '502': {description: EdgeX 명령 응답 실패 또는 결과 미확인}
        '503': {description: EdgeX 비활성화}
...
components:
  securitySchemes:
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-TOKEN
      description: 쓰기 요청은 세션 쿠키와 함께 이 헤더가 필요합니다. Swagger는 자동 발급·전달하므로 이 입력칸을 비워 두세요. 직접 호출할 때는 getCsrfToken을 먼저 사용합니다.
...
  responses:
...
    VirtualDeviceError:
      description: 잘못된 입력400·없는 VD/Profile/Device/Node404·중복 키/revision/원본 충돌409·64KiB 초과413·DB 장애503입니다. code와 message를 확인하세요.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiError'}
...
  schemas:
    SensorRegistrationTemplate:
      type: object
      required: [id, label, serviceName, nodeName, profileName, protocol, resources, endpoint, baudRates]
      properties:
        id: {type: string}
        label: {type: string}
        serviceName: {type: string}
        nodeName: {type: string}
        profileName: {type: string}
        protocol: {type: string, enum: [SERIAL, I2C]}
        resources: {type: array, items: {type: string}}
        endpoint: {type: string}
        baudRates: {type: array, items: {type: integer, enum: [1200, 2400, 4800, 9600, 19200, 38400, 57600, 115200, 230400, 460800, 921600]}}
    SensorRegistrationCatalog:
      type: object
      required: [templates]
      properties:
        templates: {type: array, items: {$ref: '#/components/schemas/SensorRegistrationTemplate'}}
    SensorRegistrationRequest:
      type: object
      additionalProperties: false
      required: [name, templateId, endpoint, physicalDeviceId]
      properties:
        name: {type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$'}
        templateId: {type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$'}
        endpoint: {type: string, maxLength: 80, description: 'Serial은 /dev/edgeai/ 아래의 단일 장치명, Sense HAT는 /dev/i2c-1'}
        physicalDeviceId: {type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$'}
        baudRate: {type: [integer, 'null'], enum: [null, 1200, 2400, 4800, 9600, 19200, 38400, 57600, 115200, 230400, 460800, 921600]}
    SensorRegistrationResult:
      type: object
      required: [name, serviceName, profileName, created, adminState, operatingState]
      properties:
        name: {type: string}
        serviceName: {type: string}
        profileName: {type: string}
        created: {type: boolean}
        adminState: {type: string}
        operatingState: {type: string}
    SensorReading:
      type: object
      required: [resource, valueType, value, units, observedAt]
      properties:
        resource: {type: string}
        valueType: {type: string}
        value: {type: string, description: "원본 문자열 값. Binary는 [binary]로 표시"}
        units: {type: string}
        observedAt: {type: string, format: date-time}
    SensorReadings:
      type: object
      required: [device, fetchedAt, readings]
      properties:
        device: {type: string}
        fetchedAt: {type: string, format: date-time}
        readings: {type: array, items: {$ref: '#/components/schemas/SensorReading'}}
    SensorCommand:
      type: object
      required: [name, readable, writable, parameters]
      properties:
        name: {type: string}
        readable: {type: boolean}
        writable: {type: boolean}
        parameters:
          type: array
          items:
            type: object
            required: [resource, valueType]
            properties:
              resource: {type: string}
              valueType: {type: string}
    SensorCommands:
      type: object
      required: [device, adminState, operatingState, commands]
      properties:
        device: {type: string}
        adminState: {type: string}
        operatingState: {type: string}
        commands: {type: array, items: {$ref: '#/components/schemas/SensorCommand'}}
    SensorCommandRequest:
      type: object
      required: [device, command, method, values]
      properties:
        device: {type: string}
        command: {type: string}
        method: {type: string, enum: [GET, PUT]}
        values: {type: object, additionalProperties: {type: string, maxLength: 4096}}
    SensorCommandResult:
      type: object
      required: [device, command, method, completedAt, readings]
      properties:
        device: {type: string}
        command: {type: string}
        method: {type: string, enum: [GET, PUT]}
        completedAt: {type: string, format: date-time}
        readings: {type: array, items: {$ref: '#/components/schemas/SensorReading'}}
...
    ApiError:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: string
          enum:
          - INVALID_PROFILE
          - PROFILE_CONFLICT
          - PROFILE_NOT_FOUND
          - PROFILE_STORE_UNAVAILABLE
          - PAYLOAD_TOO_LARGE
          - INVALID_DEVICE
          - DEVICE_CONFLICT
          - DEVICE_NOT_FOUND
          - NODE_NOT_FOUND
          - NODE_NOT_READY
          - DEVICE_RELEASED
          - STALE_SESSION
          - OBSERVATION_CONFLICT
          - DEVICE_STORE_UNAVAILABLE
          - INVALID_WORKFLOW
          - WORKFLOW_NOT_FOUND
          - WORKFLOW_CONFLICT
          - RUN_NOT_FOUND
          - TASK_NOT_FOUND
          - IDEMPOTENCY_CONFLICT
          - CANNOT_CANCEL
          - STREAM_NOT_IMPLEMENTED
          - STREAM_EXECUTION_POLICY_UNSUPPORTED
          - STREAM_RECOVERY_UNSUPPORTED
          - STREAM_SOURCE_INACTIVE
          - WORKFLOW_STORE_UNAVAILABLE
          - INVALID_TASK_ID
          - RESULT_STORE_UNAVAILABLE
          - RUNTIME_DISABLED
          - REMOTE_DISABLED
          - REMOTE_PROVIDER_NOT_FOUND
          - REMOTE_CONFIGURATION_CHANGED
          - REMOTE_TELEMETRY_UNSUPPORTED
          - VD_EXECUTION_DISABLED
          - VD_NOT_READY
          - VD_SERVICE_MISMATCH
          - VD_AUTOMATIC_OFFLOAD_UNSUPPORTED
          - OFFLOAD_SOURCE_CHANGED
          - OFFLOAD_LIMIT
          - OFFLOAD_RECOVERY_UNSUPPORTED
          - OFFLOAD_TARGET_INVALID
          - OPERATION_NOT_FOUND
          - INVALID_VIRTUAL_DEVICE
          - VD_NOT_FOUND
          - VD_CONFLICT
          - VD_RELEASED
          - VD_SOURCE_INCOMPATIBLE
          - VD_STORE_UNAVAILABLE
          - DEVICE_IN_USE
        message:
          type: string
...
    DeviceAttachmentInput:
      type: object
      additionalProperties: false
      required:
      - port
      properties:
        port:
          type: string
          minLength: 1
          maxLength: 128
...
    DeviceAttachment:
      type: object
      additionalProperties: false
      required:
      - id
      - deviceId
      - nodeId
      - port
      - attachedAt
      - detachedAt
      properties:
        id:
          type: string
          format: uuid
        deviceId:
          type: string
          format: uuid
        nodeId:
          type: string
          format: uuid
        port:
          type: string
        attachedAt:
          type: string
          format: date-time
        detachedAt:
          type:
          - string
          - 'null'
          format: date-time
...
    VDOperation:
      type: object
      additionalProperties: false
      required: [id, kind, vdId, requestedRevision, sourceRuntimeId, targetRuntimeId, state, reason, createdAt, updatedAt, finishedAt]
      properties:
        id: {type: string, format: uuid}
        kind: {type: string, enum: [VD_PROVISION, VD_REPLACE, VD_DRAIN]}
        vdId: {type: string, format: uuid}
        requestedRevision: {type: integer, format: int64}
        sourceRuntimeId: {type: [string, 'null'], format: uuid}
        targetRuntimeId: {type: [string, 'null'], format: uuid}
        state: {type: string, enum: [RUNNING, SUCCEEDED, FAILED, SUPERSEDED]}
        reason: {type: [string, 'null']}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
        finishedAt: {type: [string, 'null'], format: date-time}
    VDRuntime:
      type: object
      additionalProperties: false
      required: [id, vdId, generation, requestedRevision, desiredState, observedState, ready, podUid, nodeUid, nodeName, leaseUntil, readyAt, startupDeadline, drainDeadline, failureReason, createdAt, updatedAt]
      properties:
        id: {type: string, format: uuid}
        vdId: {type: string, format: uuid}
        generation: {type: integer, format: int64, minimum: 1}
        requestedRevision: {type: integer, format: int64}
        desiredState: {type: string, enum: [RUNNING, DRAINING, STOPPED]}
        observedState: {type: string, enum: [PENDING, SUBMITTED, READY, UNREADY, TERMINATED]}
        ready: {type: boolean, description: asOf 시점 준비 상태. leaseUntil 이후에는 이 응답으로 준비됨을 표시하지 마세요.}
        podUid: {type: [string, 'null'], format: uuid}
        nodeUid: {type: [string, 'null'], format: uuid}
        nodeName: {type: [string, 'null']}
        leaseUntil: {type: [string, 'null'], format: date-time}
        readyAt: {type: [string, 'null'], format: date-time}
        startupDeadline: {type: string, format: date-time}
        drainDeadline: {type: [string, 'null'], format: date-time}
        failureReason: {type: [string, 'null']}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    VDRuntimeBinding:
      type: object
      additionalProperties: false
      required: [id, vdId, runtimeId, openedRevision, closedRevision, openedAt, closedAt]
      properties:
        id: {type: string, format: uuid}
        vdId: {type: string, format: uuid}
        runtimeId: {type: string, format: uuid}
        openedRevision: {type: integer, format: int64}
        closedRevision: {type: [integer, 'null'], format: int64}
        openedAt: {type: string, format: date-time}
        closedAt: {type: [string, 'null'], format: date-time}
    VDExecution:
      type: object
      additionalProperties: false
      required: [vdId, enabled, asOf, current, pendingOperation, runtimeHistory, runtimeHistoryTruncated, bindings, bindingsTruncated, operations, operationsTruncated]
      properties:
        vdId: {type: string, format: uuid}
        enabled: {type: boolean}
        asOf: {type: string, format: date-time}
        current:
          oneOf: [{$ref: '#/components/schemas/VDRuntime'}, {type: 'null'}]
        pendingOperation:
          oneOf: [{$ref: '#/components/schemas/VDOperation'}, {type: 'null'}]
        runtimeHistory: {type: array, maxItems: 100, items: {$ref: '#/components/schemas/VDRuntime'}}
        runtimeHistoryTruncated: {type: boolean}
        bindings: {type: array, maxItems: 100, items: {$ref: '#/components/schemas/VDRuntimeBinding'}}
        bindingsTruncated: {type: boolean}
        operations: {type: array, maxItems: 100, items: {$ref: '#/components/schemas/VDOperation'}}
        operationsTruncated: {type: boolean}
...